Merge from downstream

+++ build.sh
@@ -6,9 +6,10 @@
 #
 set -e
 
-RELEASE_VERSION="1.4.550"
+RELEASE_VERSION="1.4.568"
 TOPDIR=/tmp/orchestrator-release
 export RELEASE_VERSION TOPDIR
+export GO15VENDOREXPERIMENT=1
 
 usage() {
   echo
@@ -18,20 +19,25 @@ usage() {
   echo "-t (linux|darwin) Target OS Default:(linux)"
   echo "-a (amd64|386) Arch Default:(amd64)"
   echo "-d debug output"
+  echo "-b build only, do not generate packages"
   echo "-p build prefix Default:(/usr/local)"
+  echo "-s release subversion"
   echo
 }
 
 function precheck() {
+  local target
   local ok=0 # return err. so shell exit code
 
-  if [[ ! -x "$( which fpm )" ]]; then
-    echo "Please install fpm and ensure it is in PATH"
-    ok=1
-  fi
+  if [[ "$target" == "linux" ]]; then
+    if [[ ! -x "$( which fpm )" ]]; then
+      echo "Please install fpm and ensure it is in PATH (typically: 'gem install fpm')"
+      ok=1
+    fi
 
-  if [[ ! -x "$( which rpmbuild )" ]]; then
-    echo "rpmbuild not in PATH, rpm will not be built"
+    if [[ ! -x "$( which rpmbuild )" ]]; then
+      echo "rpmbuild not in PATH, rpm will not be built (OS/X: 'brew install rpm')"
+    fi
   fi
 
   if [[ -z "$GOPATH" ]]; then
@@ -44,6 +50,11 @@ function precheck() {
     ok=1
   fi
 
+  if [[ $(go version | egrep "go1[.][01234]") ]]; then
+    echo "go version is too low. Must use 1.5 or above"
+    ok=1
+  fi
+
   return $ok
 }
 
@@ -52,7 +63,7 @@ function setuptree() {
   prefix="$1"
 
   mkdir -p $TOPDIR
-  rm -rf $TOPDIR/*
+  rm -rf ${TOPDIR:?}/*
   b=$( mktemp -d $TOPDIR/orchestratorXXXXXX ) || return 1
   mkdir -p $b/orchestrator
   mkdir -p $b/orchestrator${prefix}/orchestrator/
@@ -69,7 +80,7 @@ function oinstall() {
   cd  $(dirname $0)
   gofmt -s -w  go/
   rsync -qa ./resources $builddir/orchestrator${prefix}/orchestrator/
-  rsync -qa ./conf/*.sample $builddir/orchestrator${prefix}/orchestrator/
+  rsync -qa ./conf/orchestrator-sample.* $builddir/orchestrator${prefix}/orchestrator/
   cp etc/init.d/orchestrator.bash $builddir/orchestrator/etc/init.d/orchestrator
   chmod +x $builddir/orchestrator/etc/init.d/orchestrator
 }
@@ -81,6 +92,8 @@ function package() {
   prefix="$3"
 
   cd $TOPDIR
+
+  echo "Release version is ${RELEASE_VERSION}"
 
   case $target in
     'linux')
@@ -94,7 +107,6 @@ function package() {
       cd $TOPDIR
       # rpm packaging -- executable only
       echo "Creating Distro cli packages"
-      cp $builddir/orchestrator${prefix}/orchestrator/orchestrator $builddir/orchestrator-cli/usr/bin
       fpm -v "${RELEASE_VERSION}" --epoch 1  -f -s dir -t rpm -n orchestrator-cli -C $builddir/orchestrator-cli --prefix=/ .
       fpm -v "${RELEASE_VERSION}" --epoch 1  -f -s dir -t deb -n orchestrator-cli -C $builddir/orchestrator-cli --prefix=/ .
       ;;
@@ -102,7 +114,6 @@ function package() {
       echo "Creating Darwin full Package"
       tar -C $builddir/orchestrator -czf $TOPDIR/orchestrator-"${RELEASE_VERSION}"-$target-$arch.tar.gz ./
       echo "Creating Darwin cli Package"
-      cp $builddir/orchestrator${prefix}/orchestrator/orchestrator $builddir/orchestrator-cli/usr/bin
       tar -C $builddir/orchestrator-cli -czf $TOPDIR/orchestrator-cli-"${RELEASE_VERSION}"-$target-$arch.tar.gz ./
       ;;
   esac
@@ -117,33 +128,41 @@ function build() {
   arch="$2"
   builddir="$3"
   prefix="$4"
-  gobuild="go build -o $builddir/orchestrator${prefix}/orchestrator/orchestrator go/cmd/orchestrator/main.go"
+  ldflags="-X main.AppVersion=${RELEASE_VERSION}"
+  echo "Building via $(go version)"
+  gobuild="go build -ldflags \"$ldflags\" -o $builddir/orchestrator${prefix}/orchestrator/orchestrator go/cmd/orchestrator/main.go"
 
   case $os in
     'linux')
-      GOOS=$os GOARCH=$arch $gobuild
+      echo "GOOS=$os GOARCH=$arch $gobuild" | bash
     ;;
     'darwin')
-      GOOS=darwin GOARCH=amd64 $gobuild
+      echo "GOOS=darwin GOARCH=amd64 $gobuild" | bash
     ;;
   esac
+  [[ $(find $builddir/orchestrator${prefix}/orchestrator/ -type f -name orchestrator) ]] &&  echo "orchestrator binary created" || (echo "Failed to generate orchestrator binary" ; exit 1)
+  cp $builddir/orchestrator${prefix}/orchestrator/orchestrator $builddir/orchestrator-cli/usr/bin && echo "binary copied to orchestrator-cli" || (echo "Failed to copy orchestrator binary to orchestrator-cli" ; exit 1)
 }
 
 function main() {
-  local target arch builddir prefix
+  local target arch builddir prefix build_only
   target="$1"
   arch="$2"
   prefix="$3"
+  build_only=$4
 
-  precheck
+  precheck "$target"
   builddir=$( setuptree "$prefix" )
   oinstall "$builddir" "$prefix"
   build "$target" "$arch" "$builddir" "$prefix"
-  package "$target" "$builddir" "$prefix"
-  # cleanup
+  [[ $? == 0 ]] || return 1
+  if [[ $build_only -eq 0 ]]; then
+    package "$target" "$builddir" "$prefix"
+  fi
 }
 
-while getopts a:t:p:dh flag; do
+build_only=0
+while getopts a:t:p:s:dbh flag; do
   case $flag in
   a)
     arch="$OPTARG"
@@ -158,8 +177,15 @@ while getopts a:t:p:dh flag; do
   d)
     debug=1
     ;;
+  b)
+    echo "Build only; no packaging"
+    build_only=1
+    ;;
   p)
     prefix="$OPTARG"
+    ;;
+  s)
+    RELEASE_VERSION="${RELEASE_VERSION}_${OPTARG}"
     ;;
   ?)
     usage
@@ -174,5 +200,6 @@ arch=${arch:-"amd64"} # default for arch is amd64
 prefix=${prefix:-"/usr/local"}
 
 [[ $debug -eq 1 ]] && set -x
-main "$target" "$arch" "$prefix"
+main "$target" "$arch" "$prefix" "$build_only"
 
+echo "orchestrator build done; exit status is $?"
